fix: LayoutStat.totalMs/maxMs 改为 Long getter
之前每次访问都 new AtomicLong()，返回的是断开的副本，
语义错误且浪费分配。改为 Long getter，和 avgMs/avgUs 统一。

+++ lib/src/main/java/com/github/donglua/fastinflater/InflateTracker.kt
@@ -65,10 +65,8 @@ object InflateTracker {
         val totalNs = AtomicLong(0)
         val maxNs = AtomicLong(0)
 
-        /** 兼容旧 API：总耗时（毫秒） */
-        val totalMs: AtomicLong get() = AtomicLong(totalNs.get() / 1_000_000)
-        val maxMs: AtomicLong get() = AtomicLong(maxNs.get() / 1_000_000)
-
+        val totalMs: Long get() = totalNs.get() / 1_000_000
+        val maxMs: Long get() = maxNs.get() / 1_000_000
         val avgMs: Long get() = if (count.get() > 0) totalNs.get() / 1_000_000 / count.get() else 0
         val avgUs: Long get() = if (count.get() > 0) totalNs.get() / 1_000 / count.get() else 0
     }
